+++ CHANGES.md
@@ -339,5 +339,6 @@
 - Oticon/Bernafon/Sonic/Philips: move legacy versions checker into main checker (please edit `config.ini` referring to [Versions](https://github.com/JediLin/Hearing-Aids-Fitting-Software-Update-Checkers/wiki/Versions) information from our Wiki if need to check previous versions)
 - Bernafon: move Custom variable versions checker into main checker, with additional instruction
 - Danavox: display adjustment
+- Starkey: move Inspire OS checker into Pro Fit checker
 - Minor term adjustment
 - Remove unused scripts